requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
cssutils==2.6.0
idna==3.4
lxml==4.9.2
premailer==3.10.0
PyYAML==6.0
requests==2.29.0
soupsieve==2.4.1
urllib3==1.26.15
yagmail==0.15.293
